Make runtime the main module
