+++ docs/index.html
@@ -277,7 +277,7 @@
     <div>
       <h4>Understandable</h4>
       <p>
-        Raj is 34 lines; 203 bytes minified.
+        Raj is 34 lines; 190 bytes minified.
         This framework can fit in your head or even a tweet.
       </p>
     </div>
